pyproject.toml
[tool.poetry]
name = "erb_test"
version = "0.1.0"
description = ""
authors = ["yaseen1998 <Yasseenkouthe@ENG.HU.EDU.JO>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
django-extensions = "^3.1.5"
Werkzeug = "^2.1.2"
pandas = "^1.4.2"
scikit-learn = "^1.1.1"
SQLAlchemy = "^1.4.37"
django-allauth = "^0.51.0"
django-environ = "^0.9.0"
django-redis = "^5.2.0"
django-redis-cache = "^3.0.1"
django-rq = "^2.5.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
